Refactor `UseFBODemo`
